@@ -18,38 +18,39 @@
   },
   "homepage": "https://github.com/AAL-Guardian/guardian-backend#readme",
   "devDependencies": {
-    "@aws-sdk/credential-provider-ini": "^3.31.0",
-    "@types/aws-lambda": "^8.10.83",
-    "@types/node": "^14.17.16",
+    "@aws-sdk/credential-provider-ini": "^3.40.0",
+    "@types/aws-lambda": "^8.10.85",
+    "@types/node": "^14.17.33",
     "@types/uuid": "^8.3.1",
+    "aws-crt": "^1.10.2",
     "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^6.3.3",
-    "serverless": "^2.59.0",
+    "fork-ts-checker-webpack-plugin": "^6.4.0",
+    "serverless": "^2.66.0",
     "serverless-dynamodb-local": "^0.2.40",
     "serverless-iam-roles-per-function": "^3.2.0",
-    "serverless-offline": "^8.0.0",
+    "serverless-offline": "^8.2.0",
     "serverless-plugin-log-retention": "^2.0.0",
     "serverless-s3-local": "^0.6.20",
-    "serverless-webpack": "^5.5.4",
-    "ts-loader": "^9.2.5",
-    "typescript": "^4.4.3",
-    "webpack": "^5.52.1",
-    "webpack-cli": "^4.8.0"
+    "serverless-webpack": "^5.5.5",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.4.4",
+    "webpack": "^5.62.2",
+    "webpack-cli": "^4.9.1"
   },
   "dependencies": {
-    "@aws-sdk/client-cloudwatch-events": "^3.31.0",
-    "@aws-sdk/client-dynamodb": "^3.31.0",
-    "@aws-sdk/client-iot": "^3.31.0",
-    "@aws-sdk/client-iot-data-plane": "^3.31.0",
-    "@aws-sdk/client-lambda": "^3.31.0",
-    "@aws-sdk/client-polly": "^3.31.0",
-    "@aws-sdk/client-rds-data": "^3.31.0",
-    "@aws-sdk/client-s3": "^3.31.0",
-    "@aws-sdk/client-secrets-manager": "^3.33.0",
-    "@aws-sdk/polly-request-presigner": "^3.31.0",
-    "@aws-sdk/s3-request-presigner": "^3.31.0",
+    "@aws-sdk/client-cloudwatch-events": "^3.40.0",
+    "@aws-sdk/client-dynamodb": "^3.40.0",
+    "@aws-sdk/client-iot": "^3.40.0",
+    "@aws-sdk/client-iot-data-plane": "^3.40.0",
+    "@aws-sdk/client-lambda": "^3.40.0",
+    "@aws-sdk/client-polly": "^3.40.0",
+    "@aws-sdk/client-rds-data": "^3.40.0",
+    "@aws-sdk/client-s3": "^3.40.1",
+    "@aws-sdk/client-secrets-manager": "^3.40.0",
+    "@aws-sdk/polly-request-presigner": "^3.40.0",
+    "@aws-sdk/s3-request-presigner": "^3.40.1",
     "dayjs": "^1.10.7",
-    "mysql2": "^2.3.0",
+    "mysql2": "^2.3.2",
     "shell-exec": "^1.0.2",
     "uuid": "^8.3.2"
   }
