@@ -1,26 +1,26 @@
 
-import { DATETIME_FORMAT, executeStatement, selectStatement } from "../data/dao";
-import { ReportRequest } from "../data/models/report-request.model";
-import { getRobotAssignment, getRobotBySN } from "../data/robot";
-import { getReportRequestById, MIN_WINDOW_DIFF } from "../data/schedule";
-import { sendListenCommand, sendSpeakCommand } from "../iot/robot-commands";
-import { sendReportRequest } from "../iot/senior-app-commands";
-import dayjs = require("dayjs");
+import { selectStatement } from "../data/dao";
+import { checkUserPresence } from "../data/log-event";
 import { Client } from "../data/models/client.model";
 import { Person } from "../data/models/person.model";
+import { ReportRequest } from "../data/models/report-request.model";
+import { getRobotAssignment, getRobotBySN } from "../data/robot";
+import { getPendingReportRequest, getReportRequestById } from "../data/schedule";
+import { sendListenCommand, sendSpeakCommand } from "../iot/robot-commands";
+import { sendReportRequest } from "../iot/senior-app-commands";
 
 export async function checkUserAndLaunchReportRequest(id: string) {
   //grab report request
   const report_request = await getReportRequestById(id);
-  if(!report_request) {
+  if (!report_request) {
     console.log('Report was not found, wont launch', report_request);
     return;
   }
-  if(report_request.date_deleted) {
+  if (report_request.date_deleted) {
     console.log('Report was deleted, wont launch', report_request);
     return;
   }
-  if(report_request.date_shown) {
+  if (report_request.date_shown) {
     console.log('Report aready shown, wont launch', report_request);
     return;
   }
@@ -30,29 +30,11 @@ export async function checkUserAndLaunchReportRequest(id: string) {
     console.warn('No assigment found for report_request: ' + id);
     return;
   }
-  //check if there was voice in the last minutes
-  const guardian_log = await executeStatement(`SELECT *
-  FROM guardian_event
-  WHERE event_name = 'voice_detected'
-  AND robot_serial_number = :robot_serial_number
-  AND timestamp >= :min_time`, [
-    {
-      name: 'robot_serial_number',
-      value: {
-        stringValue: assignment.robot_serial_number
-      }
-    },{
-      name: 'min_time',
-      value: {
-        stringValue: dayjs().subtract(MIN_WINDOW_DIFF, 'minute').format(DATETIME_FORMAT)
-      }
-    }
-  ]) as { [key: string]: any; }[];
 
-  if (guardian_log.length === 0) {
+  if (!(await checkUserPresence(assignment.robot_serial_number))) {
     console.info('No voice detected in last 15 minutes, skipping report request');
     const robot = await getRobotBySN(assignment.robot_serial_number);
-    const [ client ] = await selectStatement('clients', [
+    const [client] = await selectStatement('clients', [
       {
         name: 'id',
         value: {
@@ -60,7 +42,7 @@ export async function checkUserAndLaunchReportRequest(id: string) {
         }
       }
     ]) as Client[];
-    const [ person ] = await selectStatement('persons', [
+    const [person] = await selectStatement('persons', [
       {
         name: 'id',
         value: {
@@ -69,7 +51,7 @@ export async function checkUserAndLaunchReportRequest(id: string) {
       }
     ]) as Person[]
     let message: string;
-    switch(person.language) {
+    switch (person.language) {
       case 'it':
         message = `Hey ${person.name}, sei qui?`;
         break;
@@ -85,7 +67,7 @@ export async function checkUserAndLaunchReportRequest(id: string) {
     }
     await sendSpeakCommand(robot, message, person.language);
     /** wait 3 seconds */
-    await new Promise(resolve => setTimeout(resolve, 2000) );
+    await new Promise(resolve => setTimeout(resolve, 2000));
     await sendListenCommand(robot);
     return;
   } else {
@@ -95,16 +77,27 @@ export async function checkUserAndLaunchReportRequest(id: string) {
 }
 
 export async function launchReportRequest(reportRequest: ReportRequest) {
-  if(reportRequest.date_shown) {
+  if (reportRequest.date_shown) {
     console.info('Report Request already shown, skipping');
     return;
   }
   const assignment = await getRobotAssignment(undefined, reportRequest.client_id);
   const robot = await getRobotBySN(assignment.robot_serial_number);
-  
+
   await Promise.all([
     sendReportRequest(robot, reportRequest),
   ])
-  
+}
 
+export async function checkAndLaunchPendingReports() {
+  const nextRequests = await getPendingReportRequest();
+  console.log(`Found ${nextRequests.length} Reports to elaborate`);
+  const singleRequests = nextRequests.reduce((list, curr) => {
+    if (!list.some(one => one.client_id === curr.client_id)) {
+      list.push(curr);
+    }
+    return list;
+  }, [] as ReportRequest[]);
+  console.log(`Found ${singleRequests.length} Reports really to elaborate`);
+  return await Promise.all(singleRequests.map(async one => await checkUserAndLaunchReportRequest(one.id)))
 }
